gen: Move the signals generator to a signals.rs module
